Moves all actions into routines.
