+++ libraries/python/skills/skill-library/skill_library/__init__.py
@@ -7,12 +7,11 @@ from .types import (
     ActionFn,
     AskUserFn,
     EmitFn,
-    GetStateFn,
     LanguageModel,
-    RoutineFn,
+    Metadata,
     RunContext,
     RunContextProvider,
-    SetStateFn,
+    RunRoutineFn,
 )
 from .usage import get_routine_usage
 
@@ -23,12 +22,11 @@ __all__ = [
     "EmitFn",
     "Engine",
     "get_routine_usage",
-    "GetStateFn",
     "LanguageModel",
-    "RoutineFn",
+    "Metadata",
+    "RunRoutineFn",
     "RunContext",
     "RunContextProvider",
-    "SetStateFn",
     "Skill",
     "SkillConfig",
     "SkillProtocol",
