+++ libraries/python/skills/skill-library/skill_library/skill.py
@@ -2,15 +2,29 @@
 
 import importlib
 import importlib.util
-import inspect
 import sys
 from pathlib import Path
-from typing import Protocol
+from typing import Any, Protocol, runtime_checkable
 
 from pydantic import BaseModel
 
-from .logging import logger
-from .types import ActionFn, RoutineFn
+from skill_library.types import AskUserFn, EmitFn, RunContext, RunRoutineFn
+
+from .logging import extra_data, logger
+
+
+@runtime_checkable
+class RoutineFn(Protocol):
+    async def __call__(
+        self,
+        context: RunContext,
+        routine_state: dict[str, Any],
+        emit: EmitFn,
+        run: RunRoutineFn,
+        ask_user: AskUserFn,
+        *args: Any,
+        **kwargs: Any,
+    ) -> Any: ...
 
 
 class SkillConfig(BaseModel):
@@ -26,17 +40,15 @@ class SkillProtocol(Protocol):
     """Base protocol that all skills must implement"""
 
     config: SkillConfig
-    _actions: dict[str, ActionFn]
     _routines: dict[str, RoutineFn]
 
 
 class Skill:
     def __init__(self, config: SkillConfig):
         self.config = config
-        self._actions: dict[str, ActionFn] = {}
         self._routines: dict[str, RoutineFn] = {}
 
-        # Auto-discover and register actions and routines
+        # Auto-discover and register routines.
         module = sys.modules[self.__class__.__module__]
         package_name = module.__package__ or module.__name__
         logger.info(f"Discovering skills in package: {package_name}")
@@ -51,27 +63,6 @@ class Skill:
             if not spec or not spec.origin:
                 raise ValueError(f"Could not find package path for {package_name}")
             package_path = Path(spec.origin).parent
-
-        # Look for actions (using __default__)
-        actions_path = package_path / "actions"
-        if actions_path.exists():
-            for file in actions_path.glob("*.py"):
-                if file.name == "__init__.py":
-                    continue
-                action_name = file.stem
-                action_module = importlib.import_module(f"{package_name}.actions.{action_name}")
-
-                if hasattr(action_module, "__default__"):
-                    action = action_module.__default__
-                    if callable(action) and inspect.iscoroutinefunction(action):
-                        self.register_action(action_name, action)
-                        continue
-                    else:
-                        raise ValueError(f"Action {action_name} '__default__' is not a coroutine function")
-
-                raise ValueError(f"Action module {action_name} must have a '__default__' coroutine function export")
-        else:
-            raise (ValueError(f"No actions directory found in {package_path}"))
 
         # Look for routines
         routines_path = package_path / "routines"
@@ -94,23 +85,16 @@ class Skill:
                             f"Routine {routine_name} 'main' is not a RoutineFn. Expected attributes: {routine_function_attrs}, Found: {routine_attrs}"
                         )
 
-                raise ValueError(f"Routine module {routine_name} must have a 'main' coroutine function")
-
-    def register_action(self, name: str, fn: ActionFn):
-        self._actions[name] = fn
+                logger.warning(
+                    "Routine module skipped. Routine has no `main` function.",
+                    extra_data({"routine_name": routine_name}),
+                )
 
     def register_routine(self, name: str, fn: RoutineFn):
         self._routines[name] = fn
 
-    def get_action(self, name: str) -> ActionFn | None:
-        return self._actions.get(name)
-
     def get_routine(self, name: str) -> RoutineFn | None:
         return self._routines.get(name)
-
-    def list_actions(self) -> list[str]:
-        """Return list of available action names"""
-        return list(self._actions.keys())
 
     def list_routines(self) -> list[str]:
         """Return list of available routine names"""
