--- libraries/python/skills/skill-library/skill_library/skills/common/actions/execute_research_plan.py
@@ -1,8 +1,0 @@
-from skill_library import RunContext
-
-
-async def execute_research_plan(context: RunContext):
-    pass
-
-
-__default__ = execute_research_plan
