--- libraries/python/skills/skill-library/skill_library/skills/common/actions/gather_context.py
@@ -1,8 +1,0 @@
-from skill_library import RunContext
-
-
-async def gather_context(context: RunContext):
-    pass
-
-
-__default__ = gather_context
