+++ libraries/python/skills/skill-library/skill_library/skills/common/routines/execute_research_plan.py
@@ -1,7 +1,3 @@
-"""
-web research skill
-"""
-
 from typing import Any
 
 from skill_library import AskUserFn, EmitFn, RunContext, RunRoutineFn
@@ -13,8 +9,8 @@ async def main(
     emit: EmitFn,
     run: RunRoutineFn,
     ask_user: AskUserFn,
-) -> str:
-    await ask_user("What do you want to echo?")
-    history = (await context.conversation_history()).messages
-    message = history[-1].content
-    return message
+):
+    """
+    Execute a research plan by following the steps outlined in the plan.
+    """
+    pass
