+++ libraries/python/skills/skill-library/skill_library/skills/common/routines/get_content_from_url.py
@@ -1,16 +1,16 @@
+from typing import Any
+
 import requests
 from bs4 import BeautifulSoup
-from skill_library import RunContext
-from skill_library.types import AskUserFn, EmitFn, GetStateFn, RunRoutineFn, SetStateFn
+from skill_library import AskUserFn, EmitFn, RunContext, RunRoutineFn
 
 
 async def main(
     context: RunContext,
-    ask_user: AskUserFn,
-    run: RunRoutineFn,
-    get_state: GetStateFn,
-    set_state: SetStateFn,
+    routine_state: dict[str, Any],
     emit: EmitFn,
+    run: RunRoutineFn,
+    ask_user: AskUserFn,
     url: str,
     max_length: int | None = None,
 ) -> str:
