+++ libraries/python/skills/skill-library/skill_library/skills/common/routines/gpt_complete.py
@@ -1,4 +1,4 @@
-from typing import cast
+from typing import Any, cast
 
 from openai_client import (
     CompletionError,
@@ -8,14 +8,17 @@ from openai_client import (
     message_content_from_completion,
     validate_completion,
 )
-from skill_library import RunContext
+from skill_library import AskUserFn, EmitFn, Metadata, RunContext, RunRoutineFn
 from skill_library.logging import logger
 from skill_library.skills.common import CommonSkill
-from skill_library.types import Metadata
 
 
-async def gpt_complete(
+async def main(
     context: RunContext,
+    routine_state: dict[str, Any],
+    emit: EmitFn,
+    run: RunRoutineFn,
+    ask_user: AskUserFn,
     prompt: str,
 ) -> tuple[str, Metadata]:
     """Use the vast knowledge of GPT-4 completion using any prompt you provide. All information needed for the prompt should be in the prompt. No other context or content is available from anywhere other than this prompt. Don't refer to content outside the prompt. The prompt can be big. Returns the completion."""
@@ -52,6 +55,3 @@ async def gpt_complete(
         raise completion_error from e
     else:
         return message_content_from_completion(completion), metadata
-
-
-__default__ = gpt_complete
