+++ libraries/python/skills/skill-library/skill_library/skills/posix/routines/make_home_dir.py
@@ -1,14 +1,14 @@
-from skill_library import RunContext
-from skill_library.types import AskUserFn, EmitFn, GetStateFn, RunRoutineFn, SetStateFn
+from typing import Any
+
+from skill_library import AskUserFn, EmitFn, RunContext, RunRoutineFn
 
 
 async def main(
     context: RunContext,
-    ask_user: AskUserFn,
-    run: RunRoutineFn,
-    get_state: GetStateFn,
-    set_state: SetStateFn,
+    routine_state: dict[str, Any],
     emit: EmitFn,
+    run: RunRoutineFn,
+    ask_user: AskUserFn,
     username: str,
 ) -> None:
     await run("posix.cd", "/mnt/data")
