+++ libraries/python/skills/skill-library/skill_library/skills/posix/routines/mkdir.py
@@ -1,0 +1,21 @@
+from typing import Any, cast
+
+from skill_library import AskUserFn, EmitFn, RunContext, RunRoutineFn
+from skill_library.skills.posix import PosixSkill
+
+
+async def main(
+    context: RunContext,
+    routine_state: dict[str, Any],
+    emit: EmitFn,
+    run: RunRoutineFn,
+    ask_user: AskUserFn,
+    dirname: str,
+) -> str:
+    """
+    Create a new directory.
+    """
+    posix_skill = cast(PosixSkill, context.skills["posix"])
+    shell = posix_skill.shell
+    shell.mkdir(dirname)
+    return f"Created directory {dirname}."
