+++ libraries/python/skills/skill-library/skill_library/skills/posix/routines/write_file.py
@@ -1,0 +1,22 @@
+from typing import Any, cast
+
+from skill_library import AskUserFn, EmitFn, RunContext, RunRoutineFn
+from skill_library.skills.posix.posix_skill import PosixSkill
+
+
+async def main(
+    context: RunContext,
+    routine_state: dict[str, Any],
+    emit: EmitFn,
+    run: RunRoutineFn,
+    ask_user: AskUserFn,
+    filename: str,
+    content: str,
+) -> str:
+    """
+    Write content to a file.
+    """
+    posix_skill = cast(PosixSkill, context.skills["posix"])
+    shell = posix_skill.shell
+    shell.write_file(filename, content)
+    return f"Wrote content to {filename}."
